remove spacing around the preview iframe
This is so that examples can use the whole iframe. Otherwise, if a playground has code like

```js
document.querySelector('head').innerHTML += `<style>
  html, body, #app {
    width: 100%;
    height: 100%;
    padding: 0;
  }
</style>`;
```

then an amount of content (equal to the padding) will stick beyond the bottom edge of the preview. The `block` display prevents vertical displacement due to some font configurations for inline-block elements, and we're not rendering the canvas within text.